# api-document/resources/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/is/v1/cart-service/add": {
            "post": {
                "description": "Insert into cart values ...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CART SERVICE"
                ],
                "summary": "AddCartItem",
                "operationId": "AddCartItem",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_cart_AddCartItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_cart_AddCartItemResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/cart-service/delete-cart-item": {
            "post": {
                "description": "giving the cartITemID delete the cartItem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CART SERVICE"
                ],
                "summary": "DeleteCartItem",
                "operationId": "DeleteCartItem",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_cart_DeleteCartItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_cart_DeleteCartItemResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/cart-service/delete-user-cart": {
            "post": {
                "description": "delete all the record in table cart with corresponding userID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CART SERVICE"
                ],
                "summary": "DeleteUserCart",
                "operationId": "DeleteUserCart",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_cart_DeleteUserCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_cart_DeleteUserCartResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/cart-service/get-items": {
            "post": {
                "description": "select * from cart where userID = :1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CART SERVICE"
                ],
                "summary": "GetCart",
                "operationId": "GetCart",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_cart_GetCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_cart_GetCartResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/cart-service/update": {
            "post": {
                "description": "Update the quantity of an item in the cart. front end receives ok returns then update StateQuantity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CART SERVICE"
                ],
                "summary": "UpdateCartItem",
                "operationId": "UpdateCartItem",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_cart_UpdateCartItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_cart_UpdateCartItemResponse"
                        }
                    }
                }
            }
        },
...
        "/is/v1/product-service/get-product-detail": {
            "post": {
                "description": "Get the detail of a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetProductDetail",
                "operationId": "GetProductDetail",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_product_GetProductDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_product_GetProductDetailResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/product-service/get-products": {
            "post": {
                "description": "list all the products in the inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "GetAllProduct",
                "operationId": "GetAllProduct",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_product_GetAllProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_product_GetAllProductResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/user-service/get-product-detail": {
            "post": {
                "description": "Update the user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER SERVICE"
                ],
                "summary": "UpdateUserInformation",
                "operationId": "UpdateUserInformation",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_user_UpdateUserInformationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_user_UpdateUserInformationResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/user-service/login": {
            "post": {
                "description": "login for a new session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER SERVICE"
                ],
                "summary": "Login",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_user_LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_user_LoginResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/user-service/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER SERVICE"
                ],
                "summary": "RegisterUser",
                "operationId": "RegisterUser",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_user_RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_user_RegisterResponse"
                        }
                    }
                }
            }
        },
        "/is/v1/user-service/update": {
            "post": {
                "description": "Get the user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER SERVICE"
                ],
                "summary": "GetUserInformation",
                "operationId": "GetUserInformation",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/transport.Request-github_com_kingstonduy_api-document_domain_user_GetUserInformationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/transport.Response-github_com_kingstonduy_api-document_domain_user_GetUserInformationResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_kingstonduy_api-document_domain_cart.AddCartItemRequest": {
            "type": "object",
            "properties": {
                "cartItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.AddCartItemDetail"
                    }
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_kingstonduy_api-document_domain_cart.AddCartItemResponse": {
            "type": "object"
        },
        "github_com_kingstonduy_api-document_domain_cart.DeleteCartItemRequest": {
            "type": "object",
            "properties": {
                "cartItemID": {
                    "type": "string"
                }
            }
        },
...
        "transport.Request-github_com_kingstonduy_api-document_domain_cart_AddCartItemRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.AddCartItemRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_cart_DeleteCartItemRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.DeleteCartItemRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_cart_DeleteUserCartRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.DeleteUserCartRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_cart_GetCartRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.GetCartRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_cart_UpdateCartItemRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.UpdateCartItemRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_product_GetAllProductRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetAllProductRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_product_GetProductDetailRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetProductDetailRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_user_GetUserInformationRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.GetUserInformationRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_user_LoginRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.LoginRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_user_RegisterRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.RegisterRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Request-github_com_kingstonduy_api-document_domain_user_UpdateUserInformationRequest": {
            "type": "object",
            "required": [
                "trace"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.UpdateUserInformationRequest"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Response-github_com_kingstonduy_api-document_domain_cart_AddCartItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.AddCartItemResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_cart_DeleteCartItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.DeleteCartItemResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_cart_DeleteUserCartResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.DeleteUserCartResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_cart_GetCartResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.GetCartResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_cart_UpdateCartItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_cart.UpdateCartItemResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_product_GetAllProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetAllProductResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_product_GetProductDetailResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_product.GetProductDetailResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_user_GetUserInformationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.GetUserInformationResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_user_LoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.LoginResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_user_RegisterResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.RegisterResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
        "transport.Response-github_com_kingstonduy_api-document_domain_user_UpdateUserInformationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kingstonduy_api-document_domain_user.UpdateUserInformationResponse"
                },
                "result": {
                    "$ref": "#/definitions/transport.Result"
                },
                "trace": {
                    "$ref": "#/definitions/transport.Trace"
                }
            }
        },
...
        "transport.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "transport.Trace": {
            "type": "object",
            "required": [
                "cid",
                "cts"
            ],
            "properties": {
                "cid": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "cts": {
                    "type": "integer"
                },
                "dur": {
                    "type": "integer"
                },
                "frm": {
                    "type": "string"
                },
                "replyTo": {
                    "type": "string"
                },
                "sid": {
                    "type": "string"
                },
                "sts": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "transactionTimeout": {
                    "type": "integer",
                    "maximum": 120000
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    }
}